refactor(settings): one provider setup editor, two hosts (#2772 chunk 1)
Transcribe a File's Polish step needs a key field, a model picker and an Ollama
install gate. `AIPolishSettingsView` already had all three, welded to the screen
that owned them, so the import wizard had no way to reach them — #2772 finding 9.

Splits that view along the seam it already had. SELECTION (master toggle, rail)
stays with the host, because it is per-surface: Transcribe a File picks its engine
with six cards, not a rail. SETUP moves to `ProviderSetup.swift`, because it is
identical wherever it appears.

    AIPolishSettingsView.swift   2363 -> 455 lines
    ProviderSetup.swift          new, 2044 lines

**No behaviour changes.** Everything still reads `settings.llmProvider` exactly as
it did in one view. Re-keying it onto a per-surface selection is chunk 2; doing it
here would hide a behaviour change inside a port.

FOUR PIECES, because the layout has two holes and they are not adjacent. The setup
column sits beside the rail; the Ollama catalog is full width below it. One `View`
cannot fill both, so `ProviderSetupSection` renders whichever `Part` the host asks
for and shares every helper. `ProviderSetupLifecycle` is a `ViewModifier` so each
host attaches the same five handlers to its own always-mounted container — the
detail column unmounts when AI Polish is switched off, and `onChange` still has to
see the move to `.none`. `ProviderSetupModel` is the state all of them read, and it
is deliberately dependency-free: giving it collaborators is how it would become a
second discovery or Ollama manager, which the two real ones already are.

WHAT THE MOVE PROVES, mechanically rather than by reading: the old struct declared
139 members and the two new files declare 139. The six that changed name all map —
`aiPolishKeychainUILog` -> `providerSetupKeychainUILog`, `confirmPendingOllamaDownload`
and `requestOllamaDownload` -> `ProviderSetupDownloads`, `ollamaShowsManageModels` ->
`ProviderSetupVisibility`, and `providerSelectionSurface` plus its local `selection`
binding inlined into the host body.

FOUND BY SWEEPING WHAT ELSE NAMES THE SUBJECT, which no compiler checks:
- `PolishLengthCopyTests` reads the settings view BY PATH and asserts the explainer
  copy is inside it. That copy moved, so the test would have gone red for a reason
  that had nothing to do with the copy. Repointed at the new owner.
- Three comments elsewhere cite `AIPolishSettingsView.onAppear` and
  `.egOneFailureCopy` for code that is no longer there. Repointed.
- `llm-contract.md` named `AIPolishSettingsView.isReasoningModel` as a live consumer
  of `supportsReasoning`. That symbol was deleted with the Deep-reasoning toggle in
  #1831 and swept in #2354; a grep across `Sources/` returns nothing. Pre-existing,
  unrelated to this change, corrected in place with the verification date.

SWIFT 6: `ProviderSetupVisibility` and `ProviderSetupDownloads` are `@MainActor`.
Both read `SetupCoordinator` and mutate `ProviderSetupModel`, both main-actor
isolated, and the nonisolated form was refused. The plan named this as the chunk's
open concurrency question; this is the answer.

NEW TEST, `ProviderSetupOwnershipTests`: source-spelling tripwires on the key field,
the Keychain write and the `pullModel` funnel. Verified failing against a deliberate
second `activeKeyDescriptor`, which it named. The first draft's docstring claimed the
editor "has ONE implementation"; chunk review rejected that as a claim about a PROXY —
a renamed copy, a receiver alias, or a duplicate inside `ProviderSetup.swift` itself all
escape it. The docstring now states each limit, and says what the tripwire is for anyway:
the cheap shortcut in chunk 3 is to paste a second key field into the import wizard, two
copies pass every behavioural test because both work, and no runtime assertion can see
the difference.

Plan: docs/feature-requests/issue-2772-2026-09-10-transcribe-a-file-rework.md
Brief: docs/feature-requests/chunk-1-brief.md
Self-audit-grep: /usr/bin/grep -rn "AIPolishSettingsView\." Sources/ Tests/
Self-audit-owner: Sources/EnviousWisprAppKit/Views/Settings/ProviderSetup.swift
